refactor(stores): claudeSessions.ts를 패키지로 승격 + 타입 추출
812줄 → types.ts 119줄 + index.ts 707줄. 액션 승격 없이 타입 추출만으로
한도 이내에 들어간다. create() 호출은 하나로 유지된다.

소비자 16파일의 import 경로 불변 (index.ts 가 useClaudeSessionsStore ·
SortField 재노출). 표면 스냅샷 + 기존 테스트 1,928줄 통과.

storeSurface.ts 독스트링도 함께 정정 (Task 1 리뷰 반영: "세 가지를 못
잡는다" 주장이 틀렸다는 지적을 실제 동작에 맞게 재작성).
