+++ src/dashboard/src/stores/__tests__/storeSurface.ts
@@ -1,13 +1,14 @@
 /**
  * Zustand 스토어의 공개 표면을 스냅샷으로 고정하는 헬퍼.
  *
- * `tsc --noEmit` 은 86개 소비자 파일에서 export 유실·개명을 잡지만,
- * 세 가지를 못 잡는다:
- *   1. 액션을 추출했는데 스텁이 다른 impl 에 배선된 경우 (타입은 맞음)
- *   2. 같은 타입 안에서 초기값이 바뀐 경우 (`[]` → `null`, `'desc'` → `'asc'`)
- *   3. 액션이 조용히 no-op 이 된 경우 (이름·타입 그대로)
+ * 액션을 **호출하지 않고** 이름과 초기값만 본다. 따라서 잡는 것은:
+ *   - 상태 키·액션 **이름**의 유실·추가·개명 (`tsc` 가 소비자 쪽에서 잡는 것과 중복이지만,
+ *     소비자가 함께 수정된 경우에도 남는 그물이다)
+ *   - 같은 타입 안에서 **초기값이 바뀐** 경우 (`[]` → `null`, `'desc'` → `'asc'`)
  *
- * B1·B2 의 `route_table.snapshot()` 에 해당한다.
+ * **못 잡는 것**: 스텁이 다른 이름의 impl 을 부르는 오배선, 액션이 no-op 이 된 경우.
+ * 이름·타입이 그대로라 호출 없이는 구조적으로 보이지 않는다. 그 둘은
+ * `delegation.ts` 의 위임 이름 검사가 전수로 잡는다 (Task 5 신설).
  */
 export interface StoreSurface {
   /** 함수가 아닌 상태 키. 정렬 — 선언 순서는 계약이 아니다. */
